+++ p2p/peers/peer.go
@@ -6,6 +6,7 @@ import (
 	"github.com/Qitmeer/qitmeer/common/roughtime"
 	"github.com/Qitmeer/qitmeer/core/blockdag"
 	"github.com/Qitmeer/qitmeer/core/protocol"
+	"github.com/Qitmeer/qitmeer/p2p/common"
 	pb "github.com/Qitmeer/qitmeer/p2p/proto/v1"
 	"github.com/Qitmeer/qitmeer/p2p/qnode"
 	"github.com/Qitmeer/qitmeer/p2p/qnr"
@@ -110,6 +111,21 @@ func (p *Peer) Address() ma.Multiaddr {
 	defer p.lock.RUnlock()
 
 	return p.address
+}
+
+func (p *Peer) QAddress() common.QMultiaddr {
+	p.lock.RLock()
+	defer p.lock.RUnlock()
+
+	return p.qaddress()
+}
+
+func (p *Peer) qaddress() common.QMultiaddr {
+	qma, err := common.QMultiAddrFromString(fmt.Sprintf("%s", p.address.String()+"/p2p/"+p.pid.String()))
+	if err != nil {
+		return nil
+	}
+	return qma
 }
 
 // Direction returns the direction of the given remote peer.
@@ -250,6 +266,7 @@ func (p *Peer) StatsSnapshot() (*StatsSnap, error) {
 		Direction:  p.direction,
 		GraphState: p.graphState(),
 		TimeOffset: p.timeOffset,
+		Address:    p.qaddress().String(),
 	}
 
 	n := p.node()
